+++ doc/sphinxext/numpy_ext/docscrape_sphinx.py
@@ -3,9 +3,9 @@ import inspect
 import textwrap
 import pydoc
 import sphinx
-from docscrape import NumpyDocString
-from docscrape import FunctionDoc
-from docscrape import ClassDoc
+from .docscrape import NumpyDocString
+from .docscrape import FunctionDoc
+from .docscrape import ClassDoc
 
 
 class SphinxDocString(NumpyDocString):